Start a rewrite of the program
Split the crate into a library and binary.
Attempt to minimise allocations.
